+++ App_Data/jobs/continuous/processPlaceOrderScenario/run.js
@@ -1,8 +1,4 @@
 "use strict";
-/**
- * 注文シナリオをランダムに実行し続ける
- * @ignore
- */
 var __awaiter = (this && this.__awaiter) || function (thisArg, _arguments, P, generator) {
     return new (P || (P = Promise))(function (resolve, reject) {
         function fulfilled(value) { try { step(generator.next(value)); } catch (e) { reject(e); } }
@@ -12,8 +8,12 @@ var __awaiter = (this && this.__awaiter) || function (thisArg, _arguments, P, ge
     });
 };
 Object.defineProperty(exports, "__esModule", { value: true });
+/**
+ * 注文シナリオをランダムに実行し続ける
+ */
 const sskts = require("@motionpicture/sskts-domain");
 const createDebug = require("debug");
+const mongoose = require("mongoose");
 const processPlaceOrder = require("../../../../controller/scenarios/processPlaceOrder");
 const mongooseConnectionOptions_1 = require("../../../../mongooseConnectionOptions");
 const debug = createDebug('sskts-monitoring-jobs');
@@ -23,23 +23,24 @@ if (process.env.CONTINUOUS_SCENARIOS_STOPPED === '1') {
 debug('start executing scenarios...');
 // tslint:disable-next-line:no-magic-numbers
 const INTERVAL = parseInt(process.env.CONTINUOUS_SCENARIOS_INTERVAL_IN_SECONDS, 10) * 1000;
-sskts.mongoose.connect(process.env.MONGOLAB_URI, mongooseConnectionOptions_1.default);
-const organizationRepo = new sskts.repository.Organization(sskts.mongoose.connection);
-organizationRepo.searchMovieTheaters({}).then((movieTheaters) => {
+mongoose.connect(process.env.MONGOLAB_URI, mongooseConnectionOptions_1.default);
+const sellerRepo = new sskts.repository.Seller(mongoose.connection);
+sellerRepo.search({}).then((movieTheaters) => {
     movieTheaters.forEach((movieTheater) => {
         setInterval(() => {
             // 0-{INTERVAL}の間でランダムにインターバルを置いてシナリオを実行する
             // tslint:disable-next-line:insecure-random no-magic-numbers
             const executesAfter = Math.floor(INTERVAL * Math.random());
             setTimeout(() => __awaiter(this, void 0, void 0, function* () {
+                const branchCode = movieTheater.location.branchCode;
                 try {
                     // tslint:disable-next-line:insecure-random no-magic-numbers
                     const duration = Math.floor(Math.random() * 500000 + 300000);
-                    const result = yield processPlaceOrder.main(movieTheater.location.branchCode, duration);
-                    debug('result:', result, 'movieTheater.branchCode:', movieTheater.branchCode);
+                    const result = yield processPlaceOrder.main(branchCode, duration);
+                    debug('result:', result, 'movieTheater.branchCode:', branchCode);
                 }
                 catch (error) {
-                    console.error(error, 'movieTheater.branchCode:', movieTheater.branchCode);
+                    console.error(error, 'movieTheater.branchCode:', branchCode);
                 }
             }), executesAfter);
         }, INTERVAL);