+++ App_Data/jobs/triggered/createStockTelemetry/run.ts
@@ -1,11 +1,10 @@
 /**
  * グローバル測定データを作成する
- * @ignore
  */
-
 import * as sskts from '@motionpicture/sskts-domain';
 import * as createDebug from 'debug';
 import * as moment from 'moment';
+import * as mongoose from 'mongoose';
 
 import mongooseConnectionOptions from '../../../../mongooseConnectionOptions';
 
@@ -13,13 +12,13 @@ const debug = createDebug('sskts-monitoring-jobs');
 
 export async function main() {
     debug('connecting mongodb...');
-    sskts.mongoose.connect(<string>process.env.MONGOLAB_URI, mongooseConnectionOptions);
+    mongoose.connect(<string>process.env.MONGOLAB_URI, mongooseConnectionOptions);
 
-    const organizationRepo = new sskts.repository.Organization(sskts.mongoose.connection);
-    const taskRepo = new sskts.repository.Task(sskts.mongoose.connection);
-    const telemetryRepo = new sskts.repository.Telemetry(sskts.mongoose.connection);
-    const transactionRepo = new sskts.repository.Transaction(sskts.mongoose.connection);
-    const actionRepo = new sskts.repository.Action(sskts.mongoose.connection);
+    const sellerRepo = new sskts.repository.Seller(mongoose.connection);
+    const taskRepo = new sskts.repository.Task(mongoose.connection);
+    const telemetryRepo = new sskts.repository.Telemetry(mongoose.connection);
+    const transactionRepo = new sskts.repository.Transaction(mongoose.connection);
+    const actionRepo = new sskts.repository.Action(mongoose.connection);
     debug('creating telemetry...');
 
     const dateNow = moment();
@@ -27,7 +26,7 @@ export async function main() {
     const measuredAt = moment.unix((dateNow.unix() - (dateNow.unix() % 60)));
 
     // 劇場組織ごとに販売者向け測定データを作成する
-    const movieTheaters = await organizationRepo.searchMovieTheaters({});
+    const movieTheaters = await sellerRepo.search({});
     await Promise.all(movieTheaters.map(async (movieTheater) => {
         await sskts.service.report.telemetry.createStock({
             measuredAt: measuredAt.toDate(),
@@ -50,7 +49,7 @@ export async function main() {
     });
 
     debug('diconnecting mongo...');
-    await sskts.mongoose.disconnect();
+    await mongoose.disconnect();
 }
 
 main().then(() => {