+++ App_Data/jobs/triggered/reportFlowTelemetry/run.js
@@ -2,7 +2,6 @@
 Object.defineProperty(exports, "__esModule", { value: true });
 /**
  * 測定データ報告
- * @ignore
  */
 const createDebug = require("debug");
 const Controller = require("../../../../controller/reportFlowTelemetry");