+++ CHANGELOG.md
@@ -18,6 +18,13 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
 ### Security
 
+## v5.0.1 - 2019-02-07
+
+### Changed
+
+- install @motionpicture/sskts-domain@28.0.0
+- install @motionpicture/sskts-api-nodejs-client@6.0.0
+
 ## v5.0.0 - 2019-01-30
 
 ### Removed